# HW: DMA Performance improvments 