… hardware/hdl/core/dma_aligner.vhd
… hardware/hdl/core/dma_types.vhd
… hardware/sim/testlist.sh