Host <- Host] memcpy(0x1000ef08000, 0x1000eee1000, 0x00026000)  end after 61 usec (2433.402 MB/sec)
[DDR <- Host] memcpy((nil), 0x10024e81000, 0x00026000)                   end after 47 usec (3158.245 MB/sec)
[Host <- DDR] memcpy(0x10007a31000, (nil), 0x00026000)                  end after 49 usec (3029.337 MB/sec)