@luyong6 2017-09-07
I see another example from https://github.com/ibm-capi/psl_dma
It uses 64 tags for AFU read and 32 tags for AFU write. Compared to the DMA design here, the number of read tags doubled. 

Correspondingly, it uses 128*512bit reorder buffer for read, and 64*512bit tx buffer for write. 
It can reach about 3.3GB/s 

@ThomasFuchs I think this may improve the bandwidth in the future if we really need it.